refactor: replace interfaces to utils
