+++ src/utils/interfaces/interfaces.ts
@@ -1,0 +1,7 @@
+export interface IUser {
+  id: string;
+  role?: string;
+  name: string;
+  score?: string;
+  room: string;
+}
